Добавлены JWT auth и ролевая модель

+++ src/project/api/client_routes.py
@@ -1,13 +1,19 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
 from project.schemas.user import ClientCreate, ClientSchema
 from project.core.exceptions import UserNotFound, UserAlreadyExists
-from project.api.depends import database, client_repo
+from project.api.depends import database, client_repo, get_current_client, check_for_admin_access
+from project.resource.auth import get_password_hash
 
 client_router = APIRouter()
 
 
-@client_router.get("/all_users", response_model=list[ClientSchema], status_code=status.HTTP_200_OK)
+@client_router.get(
+    "/all_users",
+    response_model=list[ClientSchema],
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_client)],
+)
 async def get_all_users() -> list[ClientSchema]:
     async with database.session() as session:
         all_users = await client_repo.get_all_users(session=session)
@@ -15,7 +21,12 @@ async def get_all_users() -> list[ClientSchema]:
     return all_users
 
 
-@client_router.get("/user/{user_id}", response_model=ClientSchema, status_code=status.HTTP_200_OK)
+@client_router.get(
+    "/user/{user_id}",
+    response_model=ClientSchema,
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_client)],
+)
 async def get_user_by_id(
     user_id: int,
 ) -> ClientSchema:
@@ -28,13 +39,20 @@ async def get_user_by_id(
     return user
 
 
-@client_router.post("/add_user", response_model=ClientSchema, status_code=status.HTTP_201_CREATED)
+@client_router.post(
+    "/add_user",
+    response_model=ClientSchema,
+    status_code=status.HTTP_201_CREATED
+)
 async def add_user(
     user_dto: ClientCreate,
+    current_client: ClientSchema = Depends(get_current_client),
 ) -> ClientSchema:
+    check_for_admin_access(client=current_client)
     try:
         async with database.session() as session:
-            new_user = await client_repo.create_user(session=session, user=user_dto)
+            user_dto.password = get_password_hash(password=user_dto.password)
+            new_user = await client_repo.create_user(session=session, client=user_dto)
     except UserAlreadyExists as error:
         raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=error.message)
 
@@ -49,13 +67,16 @@ async def add_user(
 async def update_user(
     user_id: int,
     user_dto: ClientCreate,
+    current_client: ClientSchema = Depends(get_current_client),
 ) -> ClientSchema:
+    check_for_admin_access(client=current_client)
     try:
         async with database.session() as session:
+            user_dto.password = get_password_hash(password=user_dto.password)
             updated_user = await client_repo.update_user(
                 session=session,
                 user_id=user_id,
-                user=user_dto,
+                client=user_dto,
             )
     except UserNotFound as error:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=error.message)
@@ -63,10 +84,15 @@ async def update_user(
     return updated_user
 
 
-@client_router.delete("/delete_user/{user_id}", status_code=status.HTTP_204_NO_CONTENT)
+@client_router.delete(
+    "/delete_user/{user_id}",
+    status_code=status.HTTP_204_NO_CONTENT
+)
 async def delete_user(
     user_id: int,
+    current_client: ClientSchema = Depends(get_current_client),
 ) -> None:
+    check_for_admin_access(client=current_client)
     try:
         async with database.session() as session:
             user = await client_repo.delete_user(session=session, user_id=user_id)
